Update third-party-projects.md
Meet xDiscover, the game-changer in working with Apache Spark. If you have ever worked with Apache Spark, you know it requires a deep understanding of Python, Java, or Scala, and a significant grasp of Apache Spark. That’s where xDiscover comes in, simplifying the complex process in a way that everyone can easily understand.